requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pyyaml==6.0.2
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
python-multipart==0.0.20